Made param::ran mutable
I am not sure a random number generator should be part of the parameters of the problem. I make it mutable because rngs have internal state which should not be included in Param's class invariants. In this way I can use ran even from a const Param object (asking for a random number should not modify Param's state).